updated database: integer to double

+++ database/migrations/2023_06_17_152401_create_purchases_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->foreignId("client_id")->constrained("clients")->cascadeOnDelete();
             $table->foreignId("product_id")->constrained("products")->cascadeOnDelete();
             $table->integer("quantity");
-            $table->integer("discount_unit");
-            $table->integer("total");
+            $table->double("discount_unit");
+            $table->double("total");
             $table->date("date");
             $table->timestamps();
         });
